overhaul(books): add wiki integration
Готовые книги-руководства изменены так, чтобы вместо захардкоженного текста выводить странички из игровой вики.
Добавлена возможность распечатывать тексты из вики в библиотеке.
Все /obj/machinery перемещены на layer = BELOW_OBJ_LAYER.


PR #2834

+++ code/modules/library/lib_machines.dm
@@ -67,17 +67,45 @@
 	density = 1
 
 /obj/machinery/bookbinder/attackby(obj/O as obj, mob/user as mob)
-	if(istype(O, /obj/item/weapon/paper))
-		user.drop_item()
-		O.loc = src
-		user.visible_message("[user] loads some paper into [src].", "You load some paper into [src].")
-		src.visible_message("[src] begins to hum as it warms up its printing drums.")
-		sleep(rand(200,400))
-		src.visible_message("[src] whirs as it prints and binds a new book.")
-		var/obj/item/weapon/book/b = new(src.loc)
-		b.dat += O:info
-		b.SetName("Print Job #" + "[rand(100, 999)]")
-		b.icon_state = "book[rand(1,7)]"
-		qdel(O)
+	if(operable())
+		if(istype(O, /obj/item/weapon/paper) || istype(O, /obj/item/weapon/book/wiki/template))
+			user.drop_item()
+			O.loc = src
+			user.visible_message("[user] loads some paper into [src].", "You load some paper into [src].")
+			src.visible_message("[src] begins to hum as it warms up its printing drums.")
+			sleep(rand(200,400))
+			src.visible_message("[src] whirs as it prints and binds a new book.")
+			if(istype(O, /obj/item/weapon/paper))
+				var/obj/item/weapon/paper/paper = O
+				print(paper.info, "Print Job #" + "[rand(100, 999)]")
+			if(istype(O, /obj/item/weapon/book/wiki/template))
+				var/obj/item/weapon/book/wiki/template/template = O
+				print_wiki(template.topic, template.censored)
+			qdel(O)
+		else
+			..()
 	else
 		..()
+		to_chat(user, "[src] doesn't work!")
+
+/obj/machinery/bookbinder/proc/print(text, title, author)
+	var/obj/item/weapon/book/book = new(src.loc)
+	if(text)
+		book.dat += text
+	if(title)
+		book.title = title
+		book.SetName(title)
+	if(author)
+		book.author = author
+	book.icon_state = "book[rand(1,7)]"
+	return book
+
+/obj/machinery/bookbinder/proc/print_wiki(topic, censorship)
+	var/obj/item/weapon/book/wiki/book
+	if(topic in GLOB.premade_manuals)
+		var/manual_type = GLOB.premade_manuals[topic]
+		book = new manual_type(src.loc, topic, censorship)
+	else
+		book = new /obj/item/weapon/book/wiki(src.loc, topic, censorship, WIKI_MINI)
+		book.icon_state = "book[rand(1,7)]"
+	return book
